Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> All the version-dependent offset and layout calculations in this CL scare me, even more so as a security person (even if this code is correct now, how can we make sure it remains correct?)
> 
> Can't you just use regular structs? Can use (ab)use the buffer size as a version identifier?

Hi Mattias, can you explain a bit more on how your solution works? We want to add a version byte (could be version 1, version 2...) to the key handle, but the key handle is in the middle of the existing structs, so that will change the layout - and we still want the existing key handles to work, so I'm having two sets of structs.

If we just use the existing structs -> Not sure where to encode the version byte.
If we just use a new struct that has a longer KH field -> u2fd needs to migrate existing key handles, which also feels dangerous?

Patch-set: 3
Label: Commit-Queue=0