Update patch set 3
Patch Set 3:

(7 comments)

I took a look at the FIPS requirements doc and now have a (slightly :) better idea of how this is supposed to work.

One of the concerns is error processing. Say if a in-line test detects a failure (too many zeros or ones) - should the result be discarded and another attempt to get a value be made? 

What do we do in general if there is a non-recoverable failure?

Patch-set: 3