Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 3:
> 
> My only problem at this point in this CL is the code in power/stoney.c.  I believe this code is also used by grunt and it is not accommodated for.  I will comment within that file.

I see how you made grunt work. Since this is a board level versioning issue and when the old board is no longer supported this code will no longer be needed... it would be nice if this was just hidden from the general code.

Maybe if the GPIO_S0_PGOOD stayed as it was in this code and the actual names in the zork boards that need this have GPIO_S0_PGOOD_Vx with GPIO_S0_PGOOD being defined as "enum gpio_signal" type variable. You could move the board level function over to a board file and assign the correct versioned signal to it from the choices GPIO_S0_PGOOD_Vx.  You could default it to the latest.  Then when you no longer have to support this older hardware, all of the board level code can vanish and the correct GPIO can be again called GPIO_S0_PGOOD.

If this is not a temporary workaround to deal with hardware that will vanish during the normal process and no longer have to be supported, then at least moving the board level function from power/stoney.c to the appropriate board/baseboard files will isolate to the correct location.

Patch-set: 3