TCPMv2: Set open CC's briefly on boot with no-contract debug accessory
The RAA489000 requires that the CC lines be set open between sourcing a
debug accessory and setting Rd again.  Otherwise, the TCPC gets stuck
asserting Rp to itself, causing the TC to wedge in AttachWait.SNK.

This sets open on boot specically when we've had a debug partner with
no contract (as servo_v4 doesn't run PD when sinking).

BRANCH=None
BUG=b:159495742,b:158802939
TEST=on waddledoo, firmware_FAFTSetup passes

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I3c565a2cb9254a4f16445fc20c6f2a87fda3f680