Update patch set 10
Patch Set 10:

(6 comments)

Looks good broadly.

Note that, as of https://crrev.com/c/2256620, the structure of the DP module is somewhat more elaborate to accommodate the case when we try to enter a mode, and the other side is already in that mode. We won't support TBT in recovery mode, so I don't know if the need for that in the TBT module is as acute. If your goal was to structure the TBT module similarly to the DP module, note that the DP module has now changed.

Patch-set: 10