Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> This seems bad? Wouldn't this cause a brownout without a battery?

Potentially, though I'm not seeing a brownout with my zinger and waddledoo with no battery.  I think the best plan will be to let the ODM use this CL for testing while Jett works on the partner reset.  

The debug flag turned out to not be a reliable indicator since we can get wedged before a reboot in FAFT.  Not always though. It appears that the debug flag in the TCPC is the difference between whether we initially wedge the port or not.

Patch-set: 2