TCPMv2: Remove alt mode VDM "sent" states
In alt mode modules, don't change state after constructing a VDM for the
PE to send. At that point, the VDM has not been sent, it may never be
sent due to message discard, and if it is sent, the AMS may be
interrupted. Wait to change state until the DPM notifies the alt mode
module of some response or non-response to the sent VDM.

By doing this, allow the alt mode modules to react gracefully to a
discarded message or interrupted AMS. Otherwise, they would get into an
unexpected state and give up on alt mode setup.

Reduce flash size by 104 bytes on average (volteer RO: 192; RW: 184).

BUG=b:160701588,b:158572770
TEST=Observe DP and TBT setup recover from interrupted VDM sequence
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: I3e556771f603312b1f0f574ce8d6e7884c19cf16