Update patch set 3
Patch Set 3:

if you have a spare chromebook - could you please try to use this fw and check if additional 40ms delay affects stability. I'm not sure if we need to have AP in RESET until KAT completed, so don't do it now and hope this is fine, but i only have hatch system to play with.

Patch-set: 3