Volteer: Configure LED side select PWM frequency
On volteer, there are 4 PWMs for the charging LEDs, 1 for each color
channel and 1 for the side selection.

The side selection PWM frequency needs to be 1/2 of the frequency of
color channel PWMs. This is so that each side sees every other pulse
from the color channels.

Currently it is 2X, this shows up as the C1 port flashing red and the C0
port flashing amber when the battery is almost empty and not charging.

BUG=b:139554899
TEST=With the battery near 0 and no charger connected, check that
both charging ports have the same color.
BRANCH=none

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Idf46c05c66239b02d2d71585ca950e4899d3003d