Update patch set 6
Patch Set 6:

(3 comments)

> Patch Set 6:
> 
> (3 comments)
> 
> I feel we might as well go back to a whiteboard and evaluate other solutions. I can think of at least two other solutions: fixing Cr50 (removing 2nd reset) and fixing TCPC reset (pd_partner_port_reset).

I don't think that fixing the CR50 is an option for the existing platforms like Zork and Puff, since the designs are complete and the CR50 RO is fixed.
I agree that fixing the TCP reset should fix the type-C issue (though I would like to see it in operation first to ensure that it works in all circumstances).
Though the thought here I have is that the TCPC reset doesn't fix this root problem, so I would argue that even if pd_partner_port_reset was implemented, it is still useful to have this fix so that the correct state of the EC can be relied upon - there is no guarantee that other features won't expect to be able to distinguish between power-on and reset.

Patch-set: 6