Update patch set 3
Patch Set 3: Verified-1

(3 comments)

As designed, I don't think this will work for multiple SVIDs or multiple SOPs.

This is probably a case where a short design proposal in Google Docs would be helpful.

Off the top of my head, you could have a PE flag controlling exit, and there would be accessors for the TC. The TC would initiate exit of all modes by calling the setter, and it would wait to do whatever it wants to do until the getter said exit was done. When this flag was set, the PE, in ready states, would call into the DPM to try to exit. The process would be similar to mode setup, and the DPM would maintain enough state to know how to progress the exit process. The mode-specific modules might be involved, depending on how complex exiting each mode is.

Ayushee: Can you briefly write up what you want to do, post it to the bug, and share it with the people on this CL?

Note: I'm open to accepting a CL like this that explicitly only exits the first active mode, but I don't think that would get you very far with Thunderbolt.

Patch-set: 3
Reviewer: Gerrit User 1305927 <1305927@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=-1