cr50_vboot_doc: improve picture renedering
This patch addresses gerrit markdown rendering deficiency where the
target image is not properly rendered gerrit by the gerrit server.

The images were duplicated on Google Drive, such that the smaller in
the page version of the image is still loaded from the local tree, but
the target image comes from Drive, which allows to see full size
images when required.

BUG=none
TEST=verified proper rendering of the full size images

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ief2f18bd9b3805b391b19b8ebb8828d9e5ca05b4