raa489000: Reset BFET to normal during init
When the system goes into hibernate, we disable the BFET to reduce
power.  However, we weren't returning the BFET back to normal
operation during initialization.  This was causing charging to not
work coming out of EC hibernate.

This commit simply resets the BFET to normal operation during init.

BUG=b:158695492
BRANCH=none
TEST=Build and flash waddledoo; hibernate DUT, wake via power button,
plug in AC adapter verify that DUT is charging.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Iba8aabad2036acf9dd4e3cefb3d7f937bccf8a29