Update patch set 2
Patch Set 2:

I still get the same failure on waddledoo with PS2. The TCPC used on waddledoo requires that we start sinking ASAP if VBUS is present, otherwise the board dies. This is done as a part of the init, however, it seems that with this patch, that init is being deferred somehow and is likely causing the failure.

Also, I'd expect that performing ErrorRecovery without a battery will lead to a brownout.

Patch-set: 2