main: wait on initial power on for CR50 reset
It is a known issue that CR50 will double reset the EC due to
limitations in the rbox to core hand off of the ec reset signal.

We have applied chip specific workarounds in the past. This changes
applies the workaround to all chips that opt-in.

BRANCH=none
BUG=b:151329011,b:160243292
TEST=puff only starts in RO once with power-on as reset type
TEST=applied spin_loop manually and saw wait of ~80 seconds

Apply workaround that was NPCX specific to all chips

Change-Id: I90d13adca27764be81762a7934bdeb94430456c8
Signed-off-by: Jett Rink <jettrink@chromium.org>