Update patch set 1
Patch Set 1:

I talked with Denis about this a bit and we agreed not to gate calls to the PPC driver based on the CONFIG_USB_PD_FRS_PPC. Instead, there is a check if the specific driver supports set_frs_enable in common/usbc_ppd.c.

The reason is that there can be ambiguity in who is responsible for generating the call to the PD stack when an FRS happens. Currently, when CONFIG_USB_PD_FRS_PPC is set, the PPC is responsible (volteer). If FRS_TCPC is set, the TCPC is responsible (Zork). BUT, in some sitations, such as Zork, the PPC needs configured even though the TCPC is responsible for the FRS interrupt. 

So, what we decided (and have now) is the determination of whether anything needs to be done is handled in the TCPC and PPC driver. For instance, a PPC driver could still need to do something if the TCPC handles the FRS trigger, so it would conditionally compile on USB_PD_FRS or USB_PD_FRS_TCPC. Then, since the function isn't defined, we wouldn't call it from common/usbc_ppc.c.

I don't know if this is the best way to do it but that is the background. With this change we could have ambiguity in who is responsible for the FRS trigger. 

Opinion?

Patch-set: 1