Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> Hold on, what about boards that use the PMIC to reset?

Sorry, I am not following. This is for the issue when CR50 and EC both both get power at the same time, the EC will have power and boot for a little bit until the CR50 core code and boot far enough to hold the EC in reset while it boots, then the CR50 let's the EC out of reset.

Patch-set: 3