Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I talked with Denis about this a bit and we agreed not to gate calls to the PPC driver based on the CONFIG_USB_PD_FRS_PPC. Instead, there is a check if the specific driver supports set_frs_enable in common/usbc_ppd.c.

This issue is that the driver doesn't even have a set_frs_enable member if CONFIG_USB_PD_FRS_PPC is not defined- it's omitted from usbc_pcc.h.

I suppose the config check could be pushed down into pcc_set_frs_enable, but that seems more confusing since in that configuration the function is simply a nop. Either we need to do what I've done here, or always include the set_frs_enable member in the PPC driver.

> If FRS_TCPC is set, the TCPC is responsible (Zork). BUT, in some sitations, such as Zork, the PPC needs configured even though the TCPC is responsible for the FRS interrupt.

As I understand it, the status quo is that Zork configures the PPC in board_tcpc_fast_role_swap_enable(), which is called through the TCPC configuration code path. This seems okay- even though it's configuring the PPC, this might be specific to Zork's choice of TCPC (b/138599218) combined with that particular PPC.

But it's also possible we should actually include this function in the AOZ PPC driver since it needs earlier setup than some or all TCPCs will provide. I'm unsure if TCPCs tend to specify any kind of timing for FON assertion prior to enabling the PPC, but if they don't or it's too fast for the AOZ1380 then maybe it makes sense to have that PPC driver require GPIO control for its FON signal and enable both TCPC and PPC FRS control on zork?

(Currently it looks like the FON timing for the AOZ1380 is still incorrect because the PPC signals 
overcurrent and causes port shutdown to race with FRS when I disconnect power to trigger an FRS, so
the way things are currently set up doesn't seem correct either.)

Patch-set: 1
Label: Commit-Queue=0