Update patch set 1
Patch Set 1:

> Either we need to do what I've done here, or always include the set_frs_enable member in the PPC driver.

This is the crux of the matter here, to ensure that everything still builds regardless of whether the TCPC or PPC option (or both) is enabled. The AOZ1380 question is related in that changing the driver would sidestep this current problem because we'd enable both options.

Patch-set: 1
Label: Commit-Queue=0