Waddledee: Check interrupt lines during init
The interrupt line for the charger may already be low before inits have
completed.  If this is the case, set up a deferred call to clear the
interrupts from the chips during the first run of the hooks task.

BRANCH=None
BUG=b:160618359
TEST=on waddledee, observe that when the EC reboots with the C0
interrupt line low the interrupts are processed and PD negotiation can
succeed

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Iba1b76e0b6502e1628648e08524bbc7ec1c0437b