Trogdor: Initialize the backlight PWMs to be disabled
There are 2 PWMs for the backlight: display and keyboard. The keyboard
background is controlled by the keyboard_backlight.c. The board file
doesn't need to define its behavior. So move it from the board.c.

The display backlight should be initialized to 0% duty cycle and set to
disabled at the beginning. AP will override the value later.

When resume from suspend, keep it disabled if the duty cycle is 0%.

BRANCH=None
BUG=b:157489286
TEST=When EC reboot, the display backlight (channel:1) is disabled.
> pwm
PWM channels:
  0: disabled
  1: disabled
When booting into the userspace, the display backlight is set to 34%.
> pwm
PWM channels:
  0: disabled
  1: 34%
Manually set the brightness to 0%.
> pwm
PWM channels:
  0: disabled
  1: disabled

Change-Id: I07207969b1541326e77167eadc04899554d858dd
Signed-off-by: Wai-Hong Tam <waihong@google.com>