sm5803: Use VBUS ADCs for VBUS presence
For some reason, using the CHG_DET bit in the status register does not
always leave us with the right result for VBUS presence.  However, the
VBUS threshold interrupts seem very reliable. There could be some
timing issue at play.

This commit changes the init to use the VBUS ADC measurements to
determine whether VBUS is present or not to be consistent with the
run-time notification triggered from the VBUS threshold interrupts.

BUG=b:161269327
BRANCH=None
TEST=Build and flash waddledee, with other CL to change extpower to
consult the VBUS ADCs, verify that AC presence is notified in the
system correctly during run-time and at init.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I6e694611cf58a59f20fa6843cf145421d33d0784