cr50: Check fingerprint match gpio for certain U2F requests
For boards with fingerprint MCU, there is a GPIO line from
the fingerprint MCU that can be used to indicate that the
user physically touched the fingerprint sensor. This is useful
in checking user presence in fingerprint-based u2f requests.

For Version 1 KHs, allow the fingerprint-GPIO-based presence
to replace power button press.

BUG=b:144861739
TEST=Enable fingerprint MCU to toggle the line on fingerprint match,
     and use webauthntool --get_assertion to check it works in u2f.

Change-Id: Icda80133beff2a9defdf898fb763f2b597409d9d
Signed-off-by: Yicheng Li <yichengli@chromium.org>