battery: Expose battery_manufacture_date() as API
The newer kernels request this data.

Add the battery_manufacture_date() as a new API. But change the argument
from the year/month/day to the original packed format.

The year/month/day format is showed on the console command.

Checked the TRMs of the following batteries. They don't have any way to
query the manufacture date, so return EC_ERROR_UNIMPLEMENTED.
 * bq27541
 * bq27621_g1
 * max17055
 * mm8013

BRANCH=None
BUG=b:160784792
TEST=Checked the battery console command, on both battery present and not.

Change-Id: I1deefb64f6cc594828d6c10c42fa7107dadd7559
Signed-off-by: Wai-Hong Tam <waihong@google.com>