u2f: Correct versioned key handle offset error
To reuse u2f_origin_user_keypair(), u2f_sign extracts the
key handle bytes from versioned and non-versioned requests
according to the format. In the versioned path the code
incorrectly uses the non-versioned struct to extract the
key handle bytes, which would result in wrong private key and thus
non-verifiable signature in WebAuthn on version 1 key handles.

Both the bug and the fix do not affect the non-versioned path.

BUG=b:144861739
TEST=Added debug statements to verify that u2f_generate and u2f_sign
     arrive at the same private key. With the previous buggy code
     they get different private key.

Signed-off-by: Yicheng Li <yichengli@chromium.org>
Change-Id: If79daeff98b01d050fcdc8dd69c809c184e6abb3