u2f: Support version 2 key handles (with authorization secret)
This is a proof of concept of the future u2f key handle with
authorization secret. It is working but should not be submitted.

For version 2 key handles, an authorization secret is generated
in u2f_generate and encrypted under the public key provided by
u2fd. In u2f_sign, if u2fd has the private key (from auth-time
secrets), then u2fd can provide the decrypted authorization secret
to waive power button press.

Since this is a proof of concept, it uses a dumb algorithm for
ecryption/decrytion called yicheng_dumb_algo in common/rsa.c.

BUG=none
TEST=- Used a dumb auth-time user secrets;
     - u2f_generate a version 2 key handle with the auth_time_pubkey;
     - record the encrypted_authorization_secret;
     - manually decrypt the authorization_secret using the auth-time
       private key (since Yicheng's dumb algorithm is easy to compute);
     - supply the decrypted authorization_secret, observe u2f_sign
       succeeds without requiring power button press;
     - supply a wrong authorization_secret, observe u2f_sign requires
       the power button press;

Signed-off-by: Yicheng Li <yichengli@chromium.org>
Change-Id: I1dfd1d75f7db05d044a72661a0e4b8219a4e7690