npcx: add support for rom resident sections
Because the EC images are copied in full from flash to RAM, when the
code RAM size is smaller than 1/2 the flash size the EC image size is
limited to the code RAM size, leaving unused flash space.

Create a new linker section .init_rom used to store data objects that
are single use in the previously unsed flash area. Data objects can be
used at runtime by copying into RAM using the flash_read() function.

This change is tied to the NPCX flash layout, with asserts to ensure
builds fail if the CONFIG_CHIP_INIT_ROM_REGION is not defined.

BUG=b:160330682
BRANCH=none
TEST=make buildall
TEST=With debug code, use the _init_rom macro and validate the data can
be read using flash_read().
TEST=Using hex editor, verify .init_rom section located at 192K boundary
and unused bytes are filled with 0xFF.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: Ia0785798fd1938ad6a1c254a070b219027ee82a3