baseboard/kukui: check stable_current by the timestamp
We should check the stable_current by the timestamp
rather than the inited value or not (i.e.
CHARGE_CURRENT_UNINITIALIZED). This is because
the stable_current might be modified to a previous measured
value to let the PDO selection pick the correct PDO by power
consumed. e.g.
 1. Pick 9V PDO, set stable timestamp expired for 5 mins,
    rewrite stable_current to CHARGE_CURRENT_UNINITIALIZED
 2. At next iteration, if we only check stable_current, we
    would try to pick a new PDO again, but however, the timestamp
    is not expired yet.

BUG=b:160448099
TEST=ensure kakadu charging won't switch back to 5V immediately
     after just picking 9V charging
BRANCH=kukui

Change-Id: I2bfd49a9e8fc5c6d5f4b84eb7da4262d2938f7a6
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>