tcpci: Cache value of CC_STATUS
TCPCI driver was using a mix of polling and Alerts for reading the
CC_STATUS. Cache the CC_STATUS register value so that it is only updated
by Alerts.

This change also ignores updates to CC_STATUS if Looking4Connection is
set.

BUG=none
BRANCH=none
TEST=make buildall
TEST=Verify connection of PD charger, non-PD charger, SNK against
TUSB422 and PS8815 TCPCs (volteer).
TEST=Verify connection of PD charger, non-PD charger, SNK against
TUSB422 and PS8815 TCPCs (volteer_tcpmv1).

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I9fb0e771318eedbcf07fd97db62789b5add19a57