signer: modify manifest tag field to include board name
Personalization infrastructure expects the image header tag field to
be set to the board name in ASCII with the following modifications:
the underscore in the board name replaced with a space, and the part
of the name following the underscore capitalized.

This patch modifies the chip g makefile to paste the modified board
name into the manifest, the signer copies the value into the image
header.

BUG=b:161498484
TEST=verified that the manifest is updated as expected:
  $ diff util/signer/ec_RW-manifest-dev.json /tmp/h1.signer.7YZnbg
  33c33
  <   "tag": "00000000000000000000000000000000000000000000000000000000",
  ---
  >   "tag": "68736c74204432430a00000000000000000000000000000000000000",

 and observed the personalizer test harness to retrieve the expected
 board name from the running image.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I9ecf009e21c2ab77b03c9de1ebb176197923e6e1