SM5803: Enable Vbus monitor when sourcing
The Vbus monitor is required when sourcing out power from the charger.
Without it, the charger will not output 5V to a port partner.  However,
it should still be disabled when the charger is sinking to avoid false
PROCHOT triggers.

BRANCH=None
BUG=b:160995584
TEST=on waddledee, confirm 5V can source out to Apple and HooToo
dongles.  Confirm on unplug that Vbus monitor is disabled through i2c
read.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I63ab87cc7080485a92c2e708bc70d1a398f87492