power_button: Wait more accurately for power button debounce
The existing logic may cause infinite wait, if the
power_button_wait_for_release() is called during debouncing, i.e.
after the interrupt is triggered, but before the debouncing function
power_button_change_deferred() is called.

The interrupt sets power_button_is_stable to 0. It makes entering the
loop to wait for a event. After debouncing, if found out it is a false
alarm, it sets power_button_is_stable back to 1 and does nothing.
The HOOK_POWER_BUTTON_CHANGE is not notified. So no way to escape the
loop. The HOOK_POWER_BUTTON_CHANGE is supposed to wake the task.

This CL keeps track when the debounce function was scheduled at,
i.e. the debounce_deferred_at timestamp. And wait accurately for
the power button debounce. That way has a chance to recheck the
conditions and break the loop.

The logic is a bit tricky. As the debounce function may be delayed,
the debounce_deferred_at is expired but the power button has not been
debonuced. In this case, the logic changes to polling, until the
debounce has really finished.

BRANCH=None
BUG=b:161652942
TEST=Tested a previously failed scenario: waked from hibernate, checked
AP auto-booting into S0, not stuck at S5S3.
TEST=Tested the normal case: hold Refresh + Power -> release Refresh,
checked staying in S5S3. Release power, checked transit to S0.

Change-Id: I207e7e7ed72202439c40b529fd34786ed6f7d9d5
Signed-off-by: Wai-Hong Tam <waihong@google.com>