power_button: Fix infinite wait for power button during debouncing
The existing logic may cause infinite wait, if the
power_button_wait_for_release() is called during debouncing, i.e.
after the interrupt is triggered, but before the debouncing function
power_button_change_deferred() is called.

The interrupt sets power_button_is_stable to 0. It makes entering the
loop to wait for a event. After debouncing, if found out it is a false
alarm, it sets power_button_is_stable back to 1 and does nothing.
The HOOK_POWER_BUTTON_CHANGE is not notified. So no way to escape the
loop. The HOOK_POWER_BUTTON_CHANGE is supposed to wake the task.

This CL fixes the issue by just waiting a bit of time if the debouncing
is happening. So it gets a way to break the loop.

BRANCH=None
BUG=b:161652942
TEST=Tested a previously failed scenario: waked from hibernate, checked
AP auto-booting into S0, not stuck at S5S3.

Change-Id: Ifeef8f9449811f31287f2f1bb118186952d962f7
Signed-off-by: Wai-Hong Tam <waihong@google.com>