usb_mux: Retain mux's power state on read
Following functions need to read the mux's current mode:
1. "usb_pd_control" host command and
2. "typec" console command
For reading the current mode, mux needs to exit the low power mode.

This commit ensures that the mux's power state is retained before and
after the read by saving the status of low power mode and re-entering
the low power mode depending on the status.

BUG=b:161367077
BRANCH=None
TEST=The retimer is held in low power mode when no device attached.

Signed-off-by: Ayushee <ayushee.shah@intel.com>
Change-Id: If63f7da816d58f7ff0a0a1d695e7e4fad2d14c07