hooks: Introduce chipset resume init and suspend complete hooks
These hooks are only enabled through a new CONFIG.

The resume init hook will be used to initialize the SPI driver,
which goes to sleep on suspend. Require to initialize
it first such that it can receive a host resume event, that
notifies the normal resume hook.

The suspend complete hook is paired with the resume init hook,
which reverts the initialization of the SPI driver.

BRANCH=None
BUG=b:148149387
TEST=make buildall -j
TEST=Build successfully on both default off and defining this CONFIG.

Change-Id: I615e2bf92c75f83a7b0ab3eded61a1ef241dbdcf
Signed-off-by: Wai-Hong Tam <waihong@google.com>